+++ lib/uni_klee_runtime.h
@@ -11,7 +11,7 @@
 
 
 int __cpr_choice(char* lid, char* typestr,
-                    int* rvals, char** rvals_ids, int rvals_size,
+                    long long* rvals, char** rvals_ids, int rvals_size,
                     int** lvals, char** lvals_ids, int lvals_size);
 
 int __cpr_output(char* id, char* typestr, int value);
